Add colour system bug
If reserving two colours for 2 aggergates seperated by an aggregate with a default colour, and the reserved colours were of a lower or equal index of the aggregate index, and the aggregates were asset/signal combinations, the unreserved aggregates would get the default colours in the wrong order in the widget body (legends would be correct).  Largely introduced and missed during test as the legends and the widget aggregate metadata (incl. colour) were calculated in two different ways.